update axi stuff
