+++ src/modules/axi/axil_slave.py
@@ -47,6 +47,7 @@ class AxiLiteSlave(Elaboratable, ABC):
                 with m.If(self.bus.read_data.ready):
                     m.next = "IDLE"
             with m.State("WRITE"):
+                m.d.comb += self.bus.write_data.ready.eq(self.write_done)
                 with m.If(self.bus.write_data.valid):
                     self.handle_write(m, addr, self.bus.write_data.data, self.bus.write_response.resp)
                     with m.If(self.write_done):
@@ -59,27 +60,33 @@ class AxiLiteSlave(Elaboratable, ABC):
 
         return m
 
+# def connect_lossy()
 
-class AxiLiteReg(AxiLiteSlave):
-    def __init__(self, *, width, base_address):
-        super().__init__()
-        assert width <= len(self.bus.read_data)
-        self.reg = Signal(width)
-        self.base_address = base_address
+class AxiLiteSlaveToFullBridge(Elaboratable):
+    def __init__(self):
+        self.lite_bus = Interface(addr_bits=32, data_bits=32, lite=True)
+        self.full_bus = Interface(addr_bits=32, data_bits=32, lite=False, id_bits=12)
 
-    def handle_read(self, m, addr, data, resp):
-        with m.If(addr == self.base_address):
-            m.d.sync += data.eq(self.reg)
-            m.d.sync += resp.eq(Response.OKAY)
-            m.d.sync += self.read_done.eq(1)
+    def elaborate(self, platform):
+        m = Module()
 
-    def handle_write(self, m, addr, data, resp):
-        with m.If(addr == self.base_address):
-            m.d.sync += self.reg.eq(data)
-            m.d.sync += resp.eq(Response.OKAY)
-            m.d.sync += self.write_done.eq(1)
+        m.d.comb += self.full_bus.connect(self.lite_bus, exclude = { "read_address": { "id": True }, "write_address": { "id": True }, "read_data": { "id": True }, "write_data": { "id": True }, "write_response": { "id": True } })
 
+        read_id = Signal.like(self.full_bus.read_data.id)
+        write_id = Signal.like(self.full_bus.write_data.id)
 
-if __name__ == "__main__":
-    pins = Signal(8)
-    print(verilog.convert(AxiLiteReg(width=8, base_address=0x00000000)))
+        with m.If(self.full_bus.read_address.valid):
+            m.d.comb += self.full_bus.read_data.id.eq(self.full_bus.read_address.id)
+            m.d.sync += read_id.eq(self.full_bus.read_address.id)
+        with m.Else():
+            m.d.comb += self.full_bus.read_data.id.eq(read_id)
+
+        with m.If(self.full_bus.write_address.valid):
+            m.d.comb += self.full_bus.write_data.id.eq(self.full_bus.write_address.id)
+            m.d.sync += write_id.eq(self.full_bus.write_address.id)
+        with m.Else():
+            m.d.comb += self.full_bus.write_data.id.eq(write_id)
+
+        m.d.comb += self.full_bus.read_data.last.eq(1)
+
+        return m
